requirements.txt
web3==7.11.1
pyyaml==6.0.2
colorama==0.4.6